+++ app/controllers/exports_controller.rb
@@ -1,5 +1,7 @@
 class ExportsController < ApplicationController
   before_action :authenticate_user!
+  before_action :set_viewable_project, only: [ :index, :show, :progress, :mark_unread, :destroy ]
+  before_action :redirect_without_project, only: [ :index, :show, :progress, :mark_unread, :destroy ]
   before_action :set_viewable_export, only: [ :show, :mark_unread, :progress ]
   before_action :set_editable_export, only: [ :destroy ]
   before_action :redirect_without_export, only: [ :show, :mark_unread, :progress, :destroy ]
@@ -12,20 +14,20 @@ class ExportsController < ApplicationController
   # GET /exports.json
   def index
     @order = scrub_order(Export, params[:order], "exports.created_at DESC")
-    @exports = current_user.all_viewable_exports.filter(params).search(params[:search]).order(@order).page(params[:page]).per( 20 )
+    @exports = current_user.all_viewable_exports.where(project_id: @project.id).filter(params).search(params[:search]).order(@order).page(params[:page]).per( 20 )
   end
 
   # GET /exports/1
   # GET /exports/1.json
   def show
-    @export.update_attribute :viewed, true if @export.status == 'ready'
+    @export.update viewed: true if @export.status == 'ready'
   end
 
   def mark_unread
-    @export.update_attribute :viewed, false
+    @export.update viewed: false
 
     respond_to do |format|
-      format.html { redirect_to exports_path }
+      format.html { redirect_to project_exports_path(@project) }
       format.json { render json: @export }
     end
   end
@@ -36,7 +38,7 @@ class ExportsController < ApplicationController
     @export.destroy
 
     respond_to do |format|
-      format.html { redirect_to exports_path }
+      format.html { redirect_to project_exports_path(@project) }
       format.json { head :no_content }
     end
   end
@@ -52,7 +54,7 @@ class ExportsController < ApplicationController
     end
 
     def redirect_without_export
-      empty_response_or_root_path(exports_path) unless @export
+      empty_response_or_root_path(project_exports_path(@project)) unless @export
     end
 
     # def export_params